requirements.txt
# Core dependencies
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6

# Database
supabase==2.0.3
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# AI and ML
anthropic==0.21.0
sentence-transformers==2.7.0
huggingface-hub==0.20.3
numpy==1.26.3

# PDF Processing
pdfplumber==0.10.3
PyPDF2==3.0.1
pillow==10.1.0

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# String Matching
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
unidecode==1.3.8

# Utilities
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0

# Component 10: RAG System Dependencies
anthropic==0.21.0
openai==1.12.0
scikit-learn==1.3.0
websockets==11.0.3
python-dateutil==2.8.2

# PDF Processing (Component 3)
pytesseract==0.3.13
pdf2image==1.17.0

# AI/ML (Component 5 & 6)
torch==2.1.2
transformers==4.36.2

# API & Web
requests==2.31.0
aiohttp==3.9.1

# Redis (for caching)
redis==5.0.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0

# Development tools
black==23.12.1
flake8==7.0.0
mypy==1.8.0

# Logging and monitoring
structlog==24.1.0
python-json-logger==2.0.7

# Additional dependencies
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
